// src/app/api/models/text-content.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MessageContent } from '../models/message-content';

/**
 * Text message content
 */
export type TextContent = MessageContent & {

/**
 * The text content of the message
 */
'text'?: string;
} & {
'$type': 'text';

/**
 * Type discriminator (always 'text' for TextContent)
 */
'$type': string;
};