// src/app/api/models/video-content.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { MessageContent } from '../models/message-content';

/**
 * Video message content
 */
export type VideoContent = MessageContent & {

/**
 * URL to the video
 */
'url'?: string;

/**
 * Original file name
 */
'fileName'?: string;

/**
 * File size in bytes
 */
'fileSize'?: number;

/**
 * MIME type of the file
 */
'mimeType'?: string;

/**
 * Video width in pixels
 */
'width'?: number;

/**
 * Video height in pixels
 */
'height'?: number;

/**
 * Duration of the video in seconds
 */
'durationSeconds'?: number;
} & {
'$type': 'video';

/**
 * Type discriminator (always 'video' for VideoContent)
 */
'$type': string;
};