// src/app/api/models/general-file-content.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */


/**
 * Generic file message content - immutable value object.
 * For document files, PDFs, and other non-media file types.
 */
export interface GeneralFileContent {

  /**
   * Discriminator (always 'file')
   */
  contentType: string;
}